authorizaion and authentication with jwt completed

+++ controllers/api/v1/posts_api.js
@@ -24,15 +24,17 @@ module.exports.destroy = async function(req,res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if(post.user == req.user.id) {
+        if(post.user == req.user.id) {
             post.remove();
             await Comment.deleteMany({post: req.params.id});
             return res.status(200).json({
                 message: 'Post deleted!'
             });
-        // }else {
-        //     return res.redirect('back');
-        // }
+        }else {
+            return res.status(401).json({
+                message: 'You cannot delete this post.'
+            });
+        }
     }catch(err) {
         return res.status(500).json({
             message: 'Internal Server Error'
